requirements.txt
pygame
svg.path